# api-spec/api-spec.yaml
openapi: 3.0.0
info:
  version: "1.2.3"
  title: "Unicorn API"
  description: "API for managing unicorn."
servers:
  - url: https://api-stub-server-384446500375.asia-east1.run.app/
    description: "Backend server"
components:
...
  schemas:
...
    FamilyEmail:
      type: object
      properties:
        familyEmailID:
          type: string
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        email:
          type: string
          description: "家族メールアドレス"
          example: "sample@sample.com"
        firstName:
          type: string
          description: "家族の名"
          example: "太郎"
        lastName:
          type: string
          description: "家族の姓"
          example: "山田"
        iconImageUrl:
          type: string
          description: "画像URL"
          example: https://placehold.jp/150x150.png
        phoneNumber:
          type: string
          description: "家族の電話番号"
          example: "09012345678"
    FamilyEmailRequest:
      type: object
      properties:
        email:
          type: string
          description: "家族メールアドレス"
          example: "sample@sample.com"
        firstName:
          type: string
          description: "家族の名"
          example: "太郎"
        lastName:
          type: string
          description: "家族の姓"
          example: "山田"
        iconImageUrl:
          type: string
          description: "画像URL"
          example: https://placehold.jp/150x150.png
        phoneNumber:
          type: string
          description: "家族の電話番号"
          example: "09012345678"
    FamilyEmailResponse:
      type: object
      properties:
        familyEmailID:
          type: string
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        email:
          type: string
          description: "家族メールアドレス"
          example: "sample@sample.com"
        firstName:
          type: string
          description: "家族の名"
          example: "太郎"
        lastName:
          type: string
          description: "家族の姓"
          example: "山田"
        iconImageUrl:
          type: string
          description: "画像URL"
          example: https://placehold.jp/150x150.png
        phoneNumber:
          type: string
          description: "家族の電話番号"
          example: "09012345678"
...
paths:
...
  /family_emails:
    get:
      tags:
        - family_email
      summary: "家族メールアドレス一覧取得"
      description: "家族メールアドレス一覧を取得します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
      responses:
        "200":
          description: "家族メールアドレス一覧取得成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FamilyEmail"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
    post:
      tags:
        - family_email
      summary: "家族メールアドレス登録"
      description: "家族メールアドレスを登録します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyEmailRequest"
      responses:
        "200":
          description: "家族メールアドレス登録成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyEmailResponse"
        "400":
          description: "不正なデータが送信されました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "InvalidData"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
  /family_emails/{familyEmailID}:
    put:
      tags:
        - family_email
      summary: "家族メールアドレス更新"
      description: "家族メールアドレスを更新します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
        - in: path
          name: familyEmailID
          required: true
          schema:
            type: string
            description: "家族メールアドレスUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyEmailRequest"
      responses:
        "200":
          description: "家族メールアドレス更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyEmailResponse"
        "400":
          description: "不正なデータが送信されました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "InvalidData"
        "404":
          description: "指定された家族メールアドレスが存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "FamilyEmailNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
    delete:
      tags:
        - family_email
      summary: "家族メールアドレス削除"
      description: "家族メールアドレスを削除します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
        - in: path
          name: familyEmailID
          required: true
          schema:
            type: string
            description: "家族メールアドレスUUID"
      responses:
        "204":
          description: "家族メールアドレス削除成功"
        "404":
          description: "指定された家族メールアドレスが存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "FamilyEmailNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
...